Move graphql queries inside graphql/queries.js
